Update unbound_manager.sh
Cosmetic changes.
i.e. If unbound is already installed, there is no need to display the FULL header splash box (which simply lists the install options) as this apparently confuses some users :-0.
So the new 'splash' command may be used if the user needs to see the FULL header splash box appropriate to their current mode 'Easy' or 'Advanced'

As the link to the SNB Forum unbound support thread is within the header splash box, for convenience, it is now displayed when '?'/'about' command is used.